refactor project for admin dashboard and user interface
