+++ README.md
@@ -25,7 +25,33 @@ jq -r '
     ) \(
       (.keywords // "" | strip)
     )"
-' pubmed_Data/data.json > pubmed_data/vascular_data.txt
+' pubmed_Data/vascular_data_pubmed.json > pubmed_data/vascular_data_pubmed.txt
+```
+
+now we must do the same for our jvs data, which is a bunch of pdfs in compressed format
+
+```sh
+
+# append string to data/vascular_data_jvs.txt
+```sh
+# Iterate through all compressed files in data/vascular_data_jvs
+for compressed_file in data/vascular_data_jvs/*.zip; do
+  # Unzip the compressed file into a temporary folder
+  unzip "$compressed_file" -d temp_unzipped_folder
+  # Iterate through all PDF files in the unzipped folder
+  for pdf_file in temp_unzipped_folder/*.pdf; do
+    # Extract text from the PDF file and store it in a variable
+    extracted_text=$(pdftotext "$pdf_file" -)
+    # Make the whole article be one line by replacing newlines with spaces
+    processed_text=$(echo "$extracted_text" | tr '\n' ' ')
+    # Append the processed text to the output file
+    echo "$processed_text" >> data/vascular_data_jvs.txt
+  done
+  # Clean up the temporary folder
+  rm -rf temp_unzipped_folder
+done
+```
+
 ```
 
 load env variables
@@ -42,7 +68,7 @@ poetry run python app.py
 In a separate terminal, test functionality via cURL request
 
 ```sh
-curl -X POST http://127.0.0.1:5000/query \                                            
-     -H "Content-Type: application/json" \
-     -d '{"query": "How would one treat vascular ehlers danlos syndrome?"}'
+curl -X POST http://127.0.0.1:5000/query \
+-H "Content-Type: application/json" \
+-d '{"query": "How would one treat vascular ehlers danlos syndrome?"}'
 ```
