+++ src/api/rag.py
@@ -1,7 +1,7 @@
 import os
 import sys
 from langchain_core.tools import tool
-from src.preprocess import *
+from src.api.preprocess import *
 
 @tool
 def pubmed_rag(state):
